Traincraft 1.6.4 files added to a Forge 1.7.10 project
I created a new Forge 1.7.10 project and got that to build. Then I added
in the source files for Traincraft from the 1.6.4 version and started
adding in the dependencies it's missing. The project still doesn't build
because of the errors caused between the differing Forge versions, but
hopefully now we can start work on fixing the errors and moving towards
a working 1.7.10 build.