More fixes
Packet3Chat can be replaced with sendChatMsg, less need to mess with
packets.
EntityName is now a function rather than a variable.
entityId is now ID.
addChatMessage needs a ChatComponentText(string) instead of just a
string.
renderDistance is now renderDistanceChunks
TileTrainWorkbench is an IInventory, not a chest, the code must reflect
that.
TileEntityOpenHearthFurnace is an IInventory, not a chest, code must
reflect that.
onInventoryChanged is depreciated, it was unused anyway.
if we MUST get an ID it's best to use Block.getIdFromBlock(block).
LiquidTank must be converted over to a sidedInventory.
getContainerItemStack is now getContainerItem
Entity.EntityID is now Entity.getEntityId().